requirements.txt
Django==4.2.4
django-crispy-forms==2.0
django-bootstrap-modal-forms==3.0.4
django-select2==8.1.2
django-datatable-view==2.1.6
djangorestframework==3.14.0
httpie==3.2.2
django-import-export==3.2.0
crispy-bootstrap4==2022.1
black==23.7.0

